Remove redundant trait
